Flatten packgae to main
