+++ blog_command_test.go
@@ -1,4 +1,4 @@
-package commands
+package main
 
 import (
 	"bytes"
@@ -6,12 +6,12 @@ import (
 	"testing"
 )
 
-func TestRunTwitter(t *testing.T) {
+func TestRunBlog(t *testing.T) {
 	outStream := new(bytes.Buffer)
 	errStream := new(bytes.Buffer)
-	target := &TwitterCommand{OutStream: outStream, ErrStream: errStream}
+	target := &BlogCommand{OutStream: outStream, ErrStream: errStream}
 
-	args := strings.Split("a-know twitter", " ")
+	args := strings.Split("a-know blog", " ")
 
 	if status := target.Run(args); status != ExitCodeOK {
 		t.Errorf("expected %d to eq %d", status, ExitCodeOK)
