+++ homepage_command_test.go
@@ -1,4 +1,4 @@
-package commands
+package main
 
 import (
 	"bytes"
@@ -6,12 +6,12 @@ import (
 	"testing"
 )
 
-func TestRunHatena(t *testing.T) {
+func TestRunHomepage(t *testing.T) {
 	outStream := new(bytes.Buffer)
 	errStream := new(bytes.Buffer)
-	target := &HatenaCommand{OutStream: outStream, ErrStream: errStream}
+	target := &HomepageCommand{OutStream: outStream, ErrStream: errStream}
 
-	args := strings.Split("a-know hatena", " ")
+	args := strings.Split("a-know homepage", " ")
 
 	if status := target.Run(args); status != ExitCodeOK {
 		t.Errorf("expected %d to eq %d", status, ExitCodeOK)
