+++ presentation_command_test.go
@@ -1,4 +1,4 @@
-package commands
+package main
 
 import (
 	"bytes"
@@ -6,12 +6,12 @@ import (
 	"testing"
 )
 
-func TestRunPhoto(t *testing.T) {
+func TestRunPresentation(t *testing.T) {
 	outStream := new(bytes.Buffer)
 	errStream := new(bytes.Buffer)
-	target := &PhotoCommand{OutStream: outStream, ErrStream: errStream}
+	target := &PresentationCommand{OutStream: outStream, ErrStream: errStream}
 
-	args := strings.Split("a-know photo", " ")
+	args := strings.Split("a-know presentation", " ")
 
 	if status := target.Run(args); status != ExitCodeOK {
 		t.Errorf("expected %d to eq %d", status, ExitCodeOK)
