+++ twitter_command_test.go
@@ -1,4 +1,4 @@
-package commands
+package main
 
 import (
 	"bytes"
@@ -6,12 +6,12 @@ import (
 	"testing"
 )
 
-func TestRunBlog(t *testing.T) {
+func TestRunTwitter(t *testing.T) {
 	outStream := new(bytes.Buffer)
 	errStream := new(bytes.Buffer)
-	target := &BlogCommand{OutStream: outStream, ErrStream: errStream}
+	target := &TwitterCommand{OutStream: outStream, ErrStream: errStream}
 
-	args := strings.Split("a-know blog", " ")
+	args := strings.Split("a-know twitter", " ")
 
 	if status := target.Run(args); status != ExitCodeOK {
 		t.Errorf("expected %d to eq %d", status, ExitCodeOK)
